Moved everything into files. Added some very basic tests.
