+++ tests/glitch/tests/test_main.py
@@ -1,0 +1,9 @@
+import unittest
+
+from glitch.main import glitch
+
+class TestGlitch(unittest.TestCase):
+    pass
+
+if __name__ == '__main__':
+    unittest.main()
