+++ tests/glitch/tests/test_plan.py
@@ -1,0 +1,9 @@
+import unittest
+
+from glitch.plan import generate_plan
+
+class TestPlan(unittest.TestCase):
+    pass
+
+if __name__ == '__main__':
+    unittest.main()
